Move event pattern tests into test_event_pattern.py
